refactor: Move helper code to separate files
The test.ts file got quite messy, being 2k+ lines long...
