+++ src/util.ts
@@ -1,0 +1,5 @@
+
+export async function sleep(milliseconds: number) {
+  return new Promise((resolve) => setTimeout(resolve, milliseconds));
+}
+
